Removed unused receiveQueue methods
The methods nextMsgID(), putChan(), deleteChan() are no longer, but they
have been moved to ordering_test.go to keep for benchmarking purposes.
