+++ ordering.go
@@ -63,22 +63,6 @@ func (m *receiveQueue) newCall(method string, maxReplies int, reply bool) (md *o
 	return
 }
 
-func (m *receiveQueue) nextMsgID() uint64 {
-	return atomic.AddUint64(&m.msgID, 1)
-}
-
-func (m *receiveQueue) putChan(id uint64, c chan *gorumsStreamResult) {
-	m.recvQMut.Lock()
-	m.recvQ[id] = c
-	m.recvQMut.Unlock()
-}
-
-func (m *receiveQueue) deleteChan(id uint64) {
-	m.recvQMut.Lock()
-	delete(m.recvQ, id)
-	m.recvQMut.Unlock()
-}
-
 func (m *receiveQueue) putResult(id uint64, result *gorumsStreamResult) {
 	m.recvQMut.RLock()
 	c, ok := m.recvQ[id]
